rename amm_resourse to amm_pair_provider
